Update Readme
I did a fresh install with the latest to test run installation and
ran into a few things to add here.

I noticed a bug with Postgres and Spring it seems. I believe the
error is unrelated to Portunus but will continue to research further.
For now I added the solution I found here.
https://github.com/darkskyapp/forecast-ruby/issues/13